third_party/WebKit/LayoutTests/editing/selection/missing-arguments.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>
var selection = getSelection();
var node = document.body;
shouldThrow("selection.addRange()");
shouldThrow("selection.getRangeAt()");
shouldThrow("selection.selectAllChildren()");
shouldThrow("selection.setBaseAndExtent()");
shouldThrow("selection.setBaseAndExtent(node)");
shouldThrow("selection.setBaseAndExtent(node, 0)");
shouldThrow("selection.setBaseAndExtent(node, 0, node)");
</script>
</body>
</html>
