third_party/blink/web_tests/platform/generic/editing/selection/missing-arguments-expected.txt

PASS selection.addRange() threw exception TypeError: Failed to execute 'addRange' on 'Selection': 1 argument required, but only 0 present..
PASS selection.getRangeAt() threw exception TypeError: Failed to execute 'getRangeAt' on 'Selection': 1 argument required, but only 0 present..
PASS selection.selectAllChildren() threw exception TypeError: Failed to execute 'selectAllChildren' on 'Selection': 1 argument required, but only 0 present..
PASS selection.setBaseAndExtent() threw exception TypeError: Failed to execute 'setBaseAndExtent' on 'Selection': 4 arguments required, but only 0 present..
PASS selection.setBaseAndExtent(node) threw exception TypeError: Failed to execute 'setBaseAndExtent' on 'Selection': 4 arguments required, but only 1 present..
PASS selection.setBaseAndExtent(node, 0) threw exception TypeError: Failed to execute 'setBaseAndExtent' on 'Selection': 4 arguments required, but only 2 present..
PASS selection.setBaseAndExtent(node, 0, node) threw exception TypeError: Failed to execute 'setBaseAndExtent' on 'Selection': 4 arguments required, but only 3 present..
PASS successfullyParsed is true

TEST COMPLETE
